Instrument management improvements (#149)
If some instrument hasn't loaded properly - say, because of a user
switching between 32-bit or 64-bit versions, or the plugin itself
missing for some reason, - the last valid instrument state will be
preserved in hope that it can be loaded correctly in the next launches.

Invalid instruments are shown as greyed out in the orchestra pit page;
they are supposed to be left alone, but also they can be either deleted,
or made great again^W^W valid again by adding a plugin node to them.